post

+++ app/Http/Controllers/TasksController.php
@@ -15,15 +15,37 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //
-        // メッセージ一覧を取得
-        $tasks = Task::all();
-
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
+        //$data = [];
+        $task = [];
+        //if (\Auth::id() === $task->user_id) {
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', [
             'tasks' => $tasks,
         ]);
+        }
+        else {
+            return view('welcome');
+        }
+            // メッセージ一覧を取得
+            //$tasks = Task::all();
+        
+        //return view('welcome', $data);
+    //}
+        //
+        
     }
+        // メッセージ一覧ビューでそれを表示
+        //return view('tasks.index', [
+        //    'tasks' => $tasks,
+        //]);
+    //}
 
     /**
      * Show the form for creating a new resource.
@@ -33,7 +55,7 @@ class TasksController extends Controller
     public function create()
     {
         //
-         $task = new Task;
+        $task = new Task;
 
         // メッセージ作成ビューを表示
         return view('tasks.create', [
@@ -49,18 +71,33 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'content' => 'required',
-            'status' => 'required|max:10',
+        //if (\Auth::id() === $task->user_id) {
+        if (\Auth::check()) {
+            $request->validate([
+                'content' => 'required',
+                'status' => 'required|max:10',
+            ]);
+        
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            'user_id' => $request->user_id,
         ]);
         //
         $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        //$task->user_id = $request->user_id;
+        //$task->content = $request->content;
+        //$task->status = $request->status;
+        //$task->save();
 
         // トップページへリダイレクトさせる
         return redirect('/');
+    
+        }
+        
+    else {
+        return view('welcome');
+    }
     }
 
     /**
@@ -71,12 +108,16 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        //
-        $task = Task::findOrFail($id);
         
+        $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.show', [
             'task' => $task,
         ]);
+        }
+        else {
+            return redirect ('/');
+        }
     }
 
     /**
@@ -89,9 +130,14 @@ class TasksController extends Controller
     {
         //
         $task = Task::findOrFail($id);
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        else {
+            return redirect ('/');
+        }
     }
 
     /**
@@ -101,6 +147,7 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    
     public function update(Request $request, $id)
     {
         $request->validate([
@@ -115,18 +162,23 @@ class TasksController extends Controller
         
         return redirect('/');
     }
-
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    
     public function destroy($id)
-    {
-        //
+    {   
         $task = Task::findOrFail($id);
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         return redirect('/');
+        //
+        //$task = Task::findOrFail($id);
+        //$task->delete();
+        //return redirect('/');
     }
 }
